cleanup: Move melodelete_1.py to melodelete.py with overwrite
